refactored content to vanilla-js module app
