+++ content-app/content-style.css
@@ -1,0 +1,30 @@
+.bamboo-bulk-btn-container {
+  padding:21px;
+  background-color: #f6f6f6;
+  text-align: center;
+  border-bottom: 1px solid #dbdbdb;
+}
+
+.bamboobulk_btn {
+  width: 100%;
+  background-color: rgb(0, 101, 142);
+  height: 32px;
+  padding: 0px;
+  margin: auto;
+  font-size: 14px;
+  border: 1px solid transparent;
+  box-shadow: 0 1px 0 0 rgba(0, 0, 0, 0.05);
+  font-weight: 800;
+  border-radius: 2px;
+  color: #fff;
+  cursor:pointer;
+}
+
+.one_day_del_btn, .one_day_add_btn {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  cursor: pointer;
+  opacity: 0.4;
+  z-index: 99;
+}
