+++ content-app/defaultEntries.js
@@ -1,0 +1,49 @@
+const defaultEntry = {
+    "id": null,
+    "note": "",
+    "projectId": null,
+    "taskId": null
+}
+
+export const defaultEntries = [
+    {
+        ...defaultEntry,
+        "trackingId": 1,
+        "start": "09:00",
+        "end": "13:00",
+    },
+    {
+        ...defaultEntry,
+        "trackingId": 2,
+        "start": "14:00",
+        "end": "18:00",
+    }
+]
+
+/* Final entries example
+{
+  "entries": [
+    {
+      "id": null,
+      "trackingId": 1,
+      "employeeId": 405,
+      "date": "2023-10-26",
+      "start": "15:00",
+      "end": "19:00",
+      "note": "",
+      "projectId": null,
+      "taskId": null
+    },
+    {
+      "id": null,
+      "trackingId": 2,
+      "employeeId": 405,
+      "date": "2023-10-26",
+      "start": "20:00",
+      "end": "24:00",
+      "projectId": null,
+      "taskId": null
+    }
+  ]
+}
+* */
