+++ extension/content.js
@@ -1,300 +1,1 @@
-/*global chrome*/
-
-const defaultEntry = {
-    "id": null,
-    "note": "",
-    "projectId": null,
-    "taskId": null
-}
-
-const _DEFAULT_ENTRIES_ = [
-    {
-        ...defaultEntry,
-        "trackingId": 1,
-        "start": "09:00",
-        "end": "13:00",
-    },
-    {
-        ...defaultEntry,
-        "trackingId": 2,
-        "start": "14:00",
-        "end": "18:00",
-    }
-]
-
-/* Final entries example
-{
-  "entries": [
-    {
-      "id": null,
-      "trackingId": 1,
-      "employeeId": 405,
-      "date": "2023-10-26",
-      "start": "15:00",
-      "end": "19:00",
-      "note": "",
-      "projectId": null,
-      "taskId": null
-    },
-    {
-      "id": null,
-      "trackingId": 2,
-      "employeeId": 405,
-      "date": "2023-10-26",
-      "start": "20:00",
-      "end": "24:00",
-      "projectId": null,
-      "taskId": null
-    }
-  ]
-}
-* */
-
-/*API - block start*/
-const bambooURL = "https://ninjaone.bamboohr.com/timesheet/clock/entries"
-const doOneDay = async (csrfToken, clockEntries) => {
-    await fetch(bambooURL, {
-        method: "POST",
-        credentials: "same-origin",
-        headers: {
-            "Content-Type": "application/json",
-            "X-Csrf-Token": csrfToken,
-        },
-        body: JSON.stringify({entries: clockEntries}),
-    });
-}
-const deleteOneDayEntries = async (csrfToken, clockEntries) => {
-    await fetch(bambooURL, {
-        method: "DELETE",
-        credentials: "same-origin",
-        headers: {
-            "Content-Type": "application/json",
-            "X-Csrf-Token": csrfToken,
-        },
-        body: JSON.stringify({entries: clockEntries}),
-    });
-}
-/*API - block end*/
-
-/*Utils - block start*/
-function delay(ms) {
-    return new Promise(resolve => setTimeout(resolve, ms));
-}
-
-const addZero = (i) => i < 10 ? "0" + i : i
-/*Utils - block end*/
-
-/*Timesheet data and parse - block start*/
-let _DAILY_DETAILS_
-let _EMPLOYEE_ID_
-let _CSRF_TOKEN_
-let _TODAY_DAY_NUM_
-
-const isTimesheetParsed = () => _DAILY_DETAILS_ && _EMPLOYEE_ID_ && _CSRF_TOKEN_
-const isDateInDailyDetails = (dateData) => _DAILY_DETAILS_.hasOwnProperty(dateData)
-const isDateContainsTimeEntries = (dateData) => !!_DAILY_DETAILS_[dateData]?.clockEntries?.length
-const isDateTodayOrPast = (dateData) => new Date(dateData).getDate() <= _TODAY_DAY_NUM_
-const isDateToday = (dateData) => new Date(dateData).getDate() === _TODAY_DAY_NUM_
-
-const extractCsrfToken = () => {
-    const tokenRegexp = /CSRF_TOKEN\s*=\s*"([^"]+)"/im
-    const collection = document.scripts;
-    for (let i = 0; i < collection.length; i++) {
-        const text = collection[i].text
-        if(!text)continue;
-        const match = text.match(tokenRegexp)
-        if(!!match?.length) return match[1]
-    }
-}
-
-const parseTimeSheetAndPopulateData = () => {
-    const timesheetJsonEl = document.getElementById("js-timesheet-data")
-    if (!timesheetJsonEl) return;
-
-    const timesheetJson = JSON.parse(timesheetJsonEl.textContent)
-    console.log(timesheetJson)
-
-    _DAILY_DETAILS_ = timesheetJson?.timesheet?.dailyDetails
-    _EMPLOYEE_ID_ = timesheetJson?.employeeId
-    _CSRF_TOKEN_ = extractCsrfToken() //window.CSRF_TOKEN
-    console.log("CSRF_TOKEN: ", _CSRF_TOKEN_)
-
-    _TODAY_DAY_NUM_ = new Date(timesheetJson?.today?.date ?? Date.now()).getDate()
-    console.log(_TODAY_DAY_NUM_)
-}
-/*Timesheet data and parse - block end*/
-
-/*Clock Entries storage and default - block start*/
-const clockEntriesStorageKey = "clockEntries"
-const getClockEntriesFromStorage = async () => {
-    const result = await chrome.storage.local.get(clockEntriesStorageKey)
-    if (window.__DEBUG__) {
-        console.log("Value currently is: ", result?.[clockEntriesStorageKey])
-    }
-    if (!result?.[clockEntriesStorageKey]) {
-        if (window.__DEBUG__) {
-            console.log("Not founded clockEntries in local extension storage")
-        }
-        return null
-    }
-    if (window.__DEBUG__) {
-        console.log("Founded in storage:", result[clockEntriesStorageKey])
-    }
-
-    return result[clockEntriesStorageKey]
-}
-
-const populateOneDayClockEntriesWithData = (entries, dateData, employeeId) => {
-    return entries.map((entry,index) => {
-        let {start, end} = entry
-        if(start === "24:00") {
-            start = "00:00"
-        }
-        if(end === "00:00") {
-            end = "24:00"
-        }
-        return {
-            "id": null,
-            "trackingId": index+1,
-            "employeeId": employeeId,
-            "date": dateData,
-            start,
-            end,
-            "note": "",
-            "projectId": null,
-            "taskId": null
-        }
-    })
-}
-
-const populateClockEntriesFromDefault = (dateData, employeeId) => {
-    return _DEFAULT_ENTRIES_.map(entry => {
-        return {...entry, employeeId, date: dateData}
-    })
-}
-/*Clock Entries storage and default - block end*/
-
-/*One Day buttons - block start*/
-const delBtn = (dateData) => {
-    const delBtn = document.createElement("div")
-    //delBtn.role = "button"
-    delBtn.className = "one_day_del_btn"
-    delBtn.title = "Delete clock entries"
-    delBtn.textContent = "❌" //"🗑" ⏰
-    delBtn.onclick = async (event) => {
-        event.stopPropagation()
-        const clockEntries = _DAILY_DETAILS_[dateData]?.clockEntries?.map(entry => entry.id)
-        if(!!clockEntries?.length){
-            await deleteOneDayEntries(_CSRF_TOKEN_, clockEntries)
-            location.reload()
-        }
-    }
-    //delBtn.dataset.dateData = dateData
-    return delBtn
-}
-
-const addOneDayTimeEntries = async (dateData) => {
-    console.log("Add:", dateData)
-    const storageEntries = await getClockEntriesFromStorage()
-    const clockEntries = !!storageEntries?.length
-        ? populateOneDayClockEntriesWithData(storageEntries, dateData, _EMPLOYEE_ID_)
-        : populateClockEntriesFromDefault(dateData, _EMPLOYEE_ID_)
-    await doOneDay(_CSRF_TOKEN_, clockEntries)
-    location.reload()
-}
-
-const addBtn = (dateData) => {
-    const addBtn = document.createElement("div")
-    //delBtn.role = "button"
-    addBtn.className = "one_day_add_btn"
-    addBtn.title = "Add clock entries"
-    addBtn.textContent = "🕔" //"⏰"
-    addBtn.onclick = () => addOneDayTimeEntries(dateData)
-    //delBtn.dataset.dateData = dateData
-    return addBtn
-}
-/*One Day buttons - block end*/
-
-const doBulk = async () => {
-    if(!isTimesheetParsed()) return;
-
-    const btn= document.getElementById("bamboobulk_btn")
-    btn.disabled = true
-    btn.textContent = "Processing..."
-
-    const storageEntries = await getClockEntriesFromStorage()
-    console.log("storageEntries:", storageEntries)
-
-    for (const oneDay of Object.values(_DAILY_DETAILS_)) {
-        //skipping Holidays and vacations
-        if(!oneDay.date || !!oneDay?.holidays?.length || !!oneDay?.timeOff?.length) continue;
-
-        const oneDayDate = new Date(oneDay.date)
-        const oneDayWeekNum = oneDayDate.getDay()
-
-        //only days below or equal today, except weekends
-        if(oneDayDate.getDate() > _TODAY_DAY_NUM_ || oneDayWeekNum === 6 || oneDayWeekNum === 0) continue;
-
-        //skip days that already have clockEntries
-        if(!!oneDay?.clockEntries?.length) continue;
-
-        const clockEntries = !!storageEntries?.length
-            ? populateOneDayClockEntriesWithData(storageEntries, oneDay.date, _EMPLOYEE_ID_)
-            : populateClockEntriesFromDefault(oneDay.date, _EMPLOYEE_ID_)
-
-        console.log(clockEntries)
-        await doOneDay(_CSRF_TOKEN_, clockEntries)
-        await delay(500)
-
-        console.log(oneDay.date)
-    }
-    location.reload()
-}
-
-const timeSheetContainer = document.querySelector(".TimesheetSummaryContainer")
-
-if(timeSheetContainer){
-    parseTimeSheetAndPopulateData()
-
-    if(isTimesheetParsed()) {
-        //bulk button
-        const bulkBtn = document.createElement("button")
-        bulkBtn.type = "button"
-        bulkBtn.id = "bamboobulk_btn"
-        bulkBtn.className = "bamboobulk_btn"
-        bulkBtn.textContent = "Bulk Time Entries"
-
-        const bambooBulkBtnContainer = document.createElement("div")
-        bambooBulkBtnContainer.id = "bamboo-bulk-btn-container"
-        bambooBulkBtnContainer.className = "bamboo-bulk-btn-container"
-        bambooBulkBtnContainer.appendChild(bulkBtn)
-
-        timeSheetContainer.prepend(bambooBulkBtnContainer)
-        bulkBtn.addEventListener("click", doBulk)
-
-        //populating timesheet days with del or add buttons
-        document.querySelectorAll(".TimesheetSlat").forEach(el => {
-            const dayDate = el.querySelector(".TimesheetSlat__dayDate")
-            if(!dayDate) return;
-
-            //generate date string in format '2023-10-31', equal to the key format in bamboo timesheet
-            try {
-                /*
-                * toISOString() produces not local result, so while testing in CET timezone always get one day lower result
-                * const dateData = new Date(dayDate.textContent + " " + new Date().getFullYear()).toISOString().substring(0, 10)
-                * */
-                const year = new Date().getFullYear()
-                const dateTimeData = new Date(dayDate.textContent + " " + year)
-                const dateData = `${year}-${addZero(dateTimeData.getMonth()+1)}-${addZero(dateTimeData.getDate())}`
-
-                if(isDateInDailyDetails(dateData) && isDateTodayOrPast(dateData)) {
-                    console.log("adding delBtn to:", dayDate.textContent, "dateData:", dateData)
-                    el.appendChild(isDateContainsTimeEntries(dateData) ? delBtn(dateData) : addBtn(dateData))
-                }
-            }catch (e) {
-                console.log(e)
-            }
-        })
-    }
-}
-
+(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const n of document.querySelectorAll('link[rel="modulepreload"]'))c(n);new MutationObserver(n=>{for(const o of n)if(o.type==="childList")for(const s of o.addedNodes)s.tagName==="LINK"&&s.rel==="modulepreload"&&c(s)}).observe(document,{childList:!0,subtree:!0});function a(n){const o={};return n.integrity&&(o.integrity=n.integrity),n.referrerPolicy&&(o.referrerPolicy=n.referrerPolicy),n.crossOrigin==="use-credentials"?o.credentials="include":n.crossOrigin==="anonymous"?o.credentials="omit":o.credentials="same-origin",o}function c(n){if(n.ep)return;n.ep=!0;const o=a(n);fetch(n.href,o)}})();let i,r,l,u;const b=()=>i&&r&&l,w=e=>i.hasOwnProperty(e),C=e=>{var t,a;return!!((a=(t=i[e])==null?void 0:t.clockEntries)!=null&&a.length)},T=e=>new Date(e).getDate()<=u,O=()=>{const e=/CSRF_TOKEN\s*=\s*"([^"]+)"/im,t=document.scripts;for(let a=0;a<t.length;a++){const c=t[a].text;if(!c)continue;const n=c.match(e);if(n!=null&&n.length)return n[1]}},S=()=>{var a,c;const e=document.getElementById("js-timesheet-data");if(!e)return;const t=JSON.parse(e.textContent);console.log(t),i=(a=t==null?void 0:t.timesheet)==null?void 0:a.dailyDetails,r=t==null?void 0:t.employeeId,l=O(),console.log("CSRF_TOKEN: ",l),u=new Date(((c=t==null?void 0:t.today)==null?void 0:c.date)??Date.now()).getDate(),console.log(u)},g={id:null,note:"",projectId:null,taskId:null},B=[{...g,trackingId:1,start:"09:00",end:"13:00"},{...g,trackingId:2,start:"14:00",end:"18:00"}],d="clockEntries",h=async()=>{const e=await chrome.storage.local.get(d);return window.__DEBUG__&&console.log("Value currently is: ",e==null?void 0:e[d]),e!=null&&e[d]?(window.__DEBUG__&&console.log("Founded in storage:",e[d]),e[d]):(window.__DEBUG__&&console.log("Not founded clockEntries in local extension storage"),null)},E=(e,t,a)=>e.map((c,n)=>{let{start:o,end:s}=c;return o==="24:00"&&(o="00:00"),s==="00:00"&&(s="24:00"),{id:null,trackingId:n+1,employeeId:a,date:t,start:o,end:s,note:"",projectId:null,taskId:null}}),k=(e,t)=>B.map(a=>({...a,employeeId:t,date:e})),D="https://ninjaone.bamboohr.com/timesheet/clock/entries",_=async(e,t)=>{await fetch(D,{method:"POST",credentials:"same-origin",headers:{"Content-Type":"application/json","X-Csrf-Token":e},body:JSON.stringify({entries:t})})},N=async(e,t)=>{await fetch(D,{method:"DELETE",credentials:"same-origin",headers:{"Content-Type":"application/json","X-Csrf-Token":e},body:JSON.stringify({entries:t})})},I=e=>new Promise(t=>setTimeout(t,e)),y=e=>e<10?"0"+e:e,x=e=>{const t=document.createElement("div");return t.className="one_day_del_btn",t.title="Delete clock entries",t.textContent="❌",t.onclick=async a=>{var n,o;a.stopPropagation();const c=(o=(n=i[e])==null?void 0:n.clockEntries)==null?void 0:o.map(s=>s.id);c!=null&&c.length&&(await N(l,c),location.reload())},t},P=async e=>{console.log("Add:",e);const t=await h(),a=t!=null&&t.length?E(t,e,r):k(e,r);await _(l,a),location.reload()},L=e=>{const t=document.createElement("div");return t.className="one_day_add_btn",t.title="Add clock entries",t.textContent="🕔",t.onclick=()=>P(e),t},j=()=>{document.querySelectorAll(".TimesheetSlat").forEach(e=>{const t=e.querySelector(".TimesheetSlat__dayDate");if(t)try{const a=new Date().getFullYear(),c=new Date(t.textContent+" "+a),n=`${a}-${y(c.getMonth()+1)}-${y(c.getDate())}`;w(n)&&T(n)&&(console.log("adding delBtn to:",t.textContent,"dateData:",n),e.appendChild(C(n)?x(n):L(n)))}catch(a){console.log(a)}})},v=async()=>{var a,c,n;if(!b())return;const e=document.getElementById("bamboobulk_btn");e.disabled=!0,e.textContent="Processing...";const t=await h();console.log("storageEntries:",t);for(const o of Object.values(i)){if(!o.date||(a=o==null?void 0:o.holidays)!=null&&a.length||(c=o==null?void 0:o.timeOff)!=null&&c.length)continue;const s=new Date(o.date),m=s.getDay();if(s.getDate()>u||m===6||m===0||(n=o==null?void 0:o.clockEntries)!=null&&n.length)continue;const f=t!=null&&t.length?E(t,o.date,r):k(o.date,r);console.log(f),await _(l,f),await I(500),console.log(o.date)}location.reload()},F=()=>{const e=document.createElement("button");return e.type="button",e.id="bamboobulk_btn",e.className="bamboobulk_btn",e.textContent="Bulk Time Entries",e.addEventListener("click",v),e},A=()=>{const e=document.createElement("div");return e.id="bamboo-bulk-btn-container",e.className="bamboo-bulk-btn-container",e.appendChild(F()),e},p=document.querySelector(".TimesheetSummaryContainer");p&&(S(),b()&&(p.prepend(A()),j()));
