+++ popup-react-app/package.json
