+++ popup-react-app/src/components/ClockEntry.jsx
